#38: replaced http-only cookie check with api call
