+++ src/components/common/ProtectedRoute.jsx
@@ -1,14 +1,12 @@
 import React from "react";
 import { Route, Redirect } from "react-router-dom";
 
-import { hasCookie } from "../../services/dummyAuth";
-
 const ProtectedRoute = ({ path, component: Component, render, ...rest }) => {
   return (
     <Route
       {...rest}
       render={(props) => {
-        if (!hasCookie("fabric-service")) {
+        if (localStorage.getItem("userStatus") !== "active") {
           return (
             <Redirect
               to={{
