+++ src/services/httpService.js
@@ -6,6 +6,12 @@ axios.defaults.withCredentials = true;
 axios.defaults.timeout = 20000;
 
 axios.interceptors.response.use(null, (error) => {
+    if (error.response && error.response.status === 401) {
+      // the user has not logged in
+      // do not toast error message.
+      return Promise.reject(error); 
+    }
+
     if (error.response && error.response.status === 403) {
       // the user has logged in but hasn't completed self-signup yet
       // do not toast error message.
